refactor(test-cli): share live-network pytest plugin logic
