+++ packages/testing/src/execution_testing/cli/pytest_commands/pytest_ini_files/pytest-execute-hive.ini
@@ -6,6 +6,7 @@ testpaths = tests/
 # Note: register new markers via src/execution_testing/cli/pytest_commands/plugins/shared/execute_fill.py
 addopts =
     -p execution_testing.cli.pytest_commands.plugins.shared.execute_fill
+    -p execution_testing.cli.pytest_commands.plugins.shared.live_client_flags
     -p execution_testing.cli.pytest_commands.plugins.execute.execute
     -p execution_testing.cli.pytest_commands.plugins.forks.forks
     -p execution_testing.cli.pytest_commands.plugins.concurrency
