+++ packages/testing/src/execution_testing/cli/pytest_commands/pytest_ini_files/pytest-execute-recover.ini
@@ -3,6 +3,7 @@ console_output_style = count
 minversion = 7.0
 python_files = *.py
 addopts =
+    -p execution_testing.cli.pytest_commands.plugins.shared.live_client_flags
     -p execution_testing.cli.pytest_commands.plugins.execute.execute_flags.execute_flags
     -p execution_testing.cli.pytest_commands.plugins.execute.rpc.remote
     -p execution_testing.cli.pytest_commands.plugins.execute.recover
