+++ packages/testing/src/execution_testing/cli/pytest_commands/pytest_ini_files/pytest-fill-stateful.ini
@@ -5,6 +5,7 @@ python_files = test_*.py
 testpaths = tests/
 addopts =
     -p execution_testing.cli.pytest_commands.plugins.shared.execute_fill
+    -p execution_testing.cli.pytest_commands.plugins.shared.live_client_flags
     -p execution_testing.cli.pytest_commands.plugins.filler.filler
     -p execution_testing.cli.pytest_commands.plugins.forks.forks
     -p execution_testing.cli.pytest_commands.plugins.execute.pre_alloc
